requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.8
Django==4.2.13
django-autocomplete-light==3.11.0
django-crispy-forms==2.3
idna==3.7
msal==1.29.0
pillow==10.4.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
sqlparse==0.5.0
typing-extensions==4.12.2
urllib3==2.2.2
